requirements.txt
bleach==6.2.0
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-CKEditor==1.0.0
Flask-WTF==1.2.2
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mysql-connector-python==9.2.0
pillow==11.1.0
python-dotenv==1.0.1
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.2.1
